Initialize mutex when copying property lists
Also some warning cleanup

+++ test/testhdf5.h
@@ -99,14 +99,14 @@
 /* Used to make certain a return value _is_ a value */
 #define VERIFY(_x, _val, where)                                                                              \
     do {                                                                                                     \
-        long __x = (long)_x, __val = (long)_val;                                                             \
+        long long __x = (long long)_x, __val = (long long)_val;                                                             \
         if (VERBOSE_HI) {                                                                                    \
             printf("   Call to routine: %15s at line %4d in %s had value "                                   \
-                   "%ld \n",                                                                                 \
+                   "%lld \n",                                                                                 \
                    (where), (int)__LINE__, __FILE__, __x);                                                   \
         }                                                                                                    \
         if ((__x) != (__val)) {                                                                              \
-            TestErrPrintf("*** UNEXPECTED VALUE from %s should be %ld, but is %ld at line %4d "              \
+            TestErrPrintf("*** UNEXPECTED VALUE from %s should be %lld, but is %lld at line %4d "              \
                           "in %s\n",                                                                         \
                           (where), __val, __x, (int)__LINE__, __FILE__);                                     \
             H5Eprint2(H5E_DEFAULT, stdout);                                                                  \
